rewrote, added descriptions of methods. wrote readme, some tests, rewrote sql file, decided problems with error 500 because of comments

+++ src/main/java/ru/skypro/homework/repository/CommentsRepository.java
@@ -3,10 +3,10 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.skypro.homework.model.Comment;
-
 import java.util.List;
 
 @Repository
-public interface CommentsRepository extends JpaRepository<Comment,Long> {
-    List<Comment> findAllByAdId(Long id);
+public interface CommentsRepository extends JpaRepository<Comment, Integer> {
+
+    List<Comment> findAllByAdId(Integer id);
 }
